Task 3: implement classes "ShapeManager" and "BaseShape"

+++ src/main/java/unisa/diem/swproject/model/BaseShape.java
@@ -3,10 +3,10 @@ package unisa.diem.swproject.model;
 import javafx.scene.canvas.GraphicsContext;
 
 public abstract class BaseShape implements Shape {
-    private double zIndex;
+    private int zIndex;
 
     @Override
-    public double getZIndex() {
+    public int getZIndex() {
         return zIndex;
     }
